+++ src/win32/Widgets/GCombo_Win.cpp
@@ -22,7 +22,7 @@ public:
 	GRect Pos;
 	
 	// Initialization data
-	int Len;
+	size_t Len;
 	bool Init;
 	int64 Value;
 	GString Name;
@@ -127,7 +127,7 @@ void GCombo::Value(int64 i)
 	{
 		if (d->Strs.Length() == 0 || i < 0)
 			i = 0;
-		if (i >= d->Strs.Length())
+		if (i >= (ssize_t)d->Strs.Length())
 			i = d->Strs.Length() - 1;
 
 		SendMessage(Handle(), CB_SETCURSEL, i, 0);
@@ -144,7 +144,7 @@ int64 GCombo::Value()
 			d->Value = r;
 			if (d->Strs.Length() == 0 || d->Value < 0)
 				d->Value = 0;
-			else if (d->Value >= d->Strs.Length())
+			else if (d->Value >= (ssize_t)d->Strs.Length())
 				d->Value = d->Strs.Length() - 1;
 		}
 	}
@@ -181,7 +181,7 @@ bool GCombo::Name(const char *n)
 char *GCombo::Name()
 {
 	if (d->Value >= 0 &&
-		d->Value < d->Strs.Length())
+		d->Value < (ssize_t)d->Strs.Length())
 	{
 		char *s = d->Strs[d->Value];
 
@@ -282,7 +282,7 @@ bool GCombo::Insert(const char *p, int Index)
 			return false;
 
 		if (Index < 0)
-			Index = d->Len;
+			Index = (int)d->Len;
 
 		LRESULT r = SendMessage(Handle(), CB_INSERTSTRING, Index, (LPARAM)n.Get());
 		if (r == CB_ERR)
@@ -358,7 +358,7 @@ int GCombo::SysOnNotify(int Msg, int Code)
 			{
 				uint64 Old = d->Value;
 				if (Value() != Old)
-					SendNotify(d->Value);
+					SendNotify((int)d->Value);
 				break;
 			}
 			case CBN_DROPDOWN:
@@ -402,8 +402,8 @@ GMessage::Result GCombo::OnEvent(GMessage *Msg)
 			// Force the mouse wheel to do something useful
 			LRESULT c = SendMessage(_View, CB_GETCURSEL, 0, 0);
 			LRESULT items = SendMessage(_View, CB_GETCOUNT, 0, 0);
-			int16 delta = (Msg->A() >> 16);
-			int new_cur = c - (delta / WHEEL_DELTA);
+			int16 delta = (int16) (Msg->A() >> 16);
+			int new_cur = (int) (c - (delta / WHEEL_DELTA));
 			SendMessage(_View, CB_SETCURSEL, limit(new_cur, 0, items-1), 0); 
 			break;
 		}
